requirements.txt
streamlit==1.21.0
tensorflow==2.11.0
numpy==1.23.5
Pillow==9.4.0
